+++ lint-staged.config.js
@@ -1,10 +1,8 @@
 module.exports = {
   '*.js': [
     'yarn lint:js --fix',
-    'git add',
   ],
   '*.{css,scss}': [
     'yarn lint:style --fix',
-    'git add',
   ],
 }